Finished implimenting raised cos filter
also realized that qamDecoder.c is basically geared for OFDM and won't
really work for QAM decoding in its current form. Time to branch out a
project to functionalize the QAM modulator/demodulator and impliment a
separate OFDM modulator/demodulator which is where I can put the guard
period logic I started on.