Got the Active state performing DFTs and generating IQ plots
It's working! Also made my first transmission over the airwaves with
this OFDM system, and it fucking worked! Even with the refidgerator
compressor noise, and the echoey reverberant room of the kitchen. It was
pretty clear that some of the subchannels were noiser than others, and
some were faded or amplified compared to the rest. Many of the channels
had a unique phase offset. These will be solved by calculating the
equalization parameters from the now decoded preamble IQ values